+++ webpack.config.js
@@ -19,7 +19,7 @@ module.exports = {
         loader: 'babel-loader',
         options: {
           presets: [
-            ['@babel/preset-env', { useBuiltIns: 'usage' }],
+            '@babel/preset-env',
             '@babel/preset-react',
           ],
         },